boost path

+++ Engine/e_Mesh.cpp
@@ -64,7 +64,7 @@ e_SceneInitData e_Mesh::ParseBinary(const char* a_InputFile)
 	AABB m_sLocalBox;
 	a_In >> m_sLocalBox;
 	a_In.Move(sizeof(e_MeshPartLight) * MAX_AREALIGHT_NUM + 8);
-#define PRINT(n, t) { a_In.Move(n * sizeof(t)); char msg[255]; msg[0] = 0; sprintf(msg, "Buf : %s, length : %d, size : %d[MB]\n", #t, (n), (n) * sizeof(t) / (1024 * 1024)); Platform::OutputDebug(msg);}
+#define PRINT(n, t) { a_In.Move(n * sizeof(t)); char msg[255]; msg[0] = 0; sprintf(msg, "Buf : %s, length : %llu, size : %llu[MB]\n", #t, size_t(n), size_t((n) * sizeof(t) / (1024 * 1024))); Platform::OutputDebug(msg);}
 	unsigned int m_uTriangleCount;
 	a_In >> m_uTriangleCount;
 	PRINT(m_uTriangleCount, e_TriangleData)
